Actualizar .gitignore, cambiar tipo de dato de balance en la entidad User y agregar scripts de configuración en index.html

+++ lib/domain/entities/users.dart
@@ -5,7 +5,7 @@ class User {
   final int role;
   final String avatar;
   final bool banned;
-  final double? balance;
+  final int? balance;
   final String calle;
   final Provincia provincia;
   final Localidad localidad;
